% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDrugUse2.R
\name{addDrugUse2}
\alias{addDrugUse2}
\title{Add new columns with drug use related information}
\usage{
addDrugUse2(
  cohort,
  indexDate = "cohort_start_date",
  censorDate = "cohort_end_date",
  ingredientConceptId = NULL,
  conceptSet = NULL,
  restrictIncident = TRUE,
  gapEra = 0
)
}
\arguments{
\item{cohort}{Cohort in the cdm}

\item{indexDate}{Name of a column that indicates the date to start the
analysis.}

\item{censorDate}{Name of a column that indicates the date to stop the
analysis, if NULL end of individuals observation is used.}

\item{ingredientConceptId}{Ingredient OMOP concept that we are interested for
the study. It is a compulsory input, no default value is provided.}

\item{conceptSet}{List of concepts to be included. If NULL all the
descendants of ingredient concept id will be used.
ingredientConceptId = NULL,}

\item{restrictIncident}{Whether to include only incident prescriptions in the
analysis. If FALSE all prescriptions that overlap with the study period will
be included.}

\item{gapEra}{Number of days between two continuous exposures to be
considered in the same era.}
}
\value{
The same cohort with the added columns.
}
\description{
Add new columns with drug use related information
}
\examples{
\donttest{
library(DrugUtilisation)
library(CodelistGenerator)

cdm <- mockDrugUtilisation()
cdm <- generateDrugUtilisationCohortSet(
  cdm, "dus_cohort", getDrugIngredientCodes(cdm, name = "acetaminophen")
)
cdm[["dus_cohort"]] \%>\%
  addDrugUse(ingredientConceptId = 1125315)
}

}
