% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{addIndication}
\alias{addIndication}
\title{Add a variable indicating individuals indications}
\usage{
addIndication(
  cohort,
  indicationCohortName,
  indicationCohortId = NULL,
  indicationWindow = list(c(0, 0)),
  unknownIndicationTable = NULL,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  mutuallyExclusive = TRUE,
  nameStyle = NULL,
  name = NULL,
  restrictIncident = FALSE
)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{indicationCohortName}{Name of the cohort table containing potential
indications.}

\item{indicationCohortId}{Cohort definition IDs of the indications of
interest. If \code{NULL}, all cohorts in \code{indicationCohortName} are included.}

\item{indicationWindow}{Time windows over which to identify indications.}

\item{unknownIndicationTable}{Tables in the OMOP CDM to search for unknown
indications.}

\item{indexDate}{Name of a column that indicates the date to start the
analysis.}

\item{censorDate}{Name of a column that indicates the date to stop the
analysis, if NULL end of individuals observation is used.}

\item{mutuallyExclusive}{Whether intersections should be mutually exclusive.
If \code{TRUE}, cohort combinations are reported as mutually exclusive categories;
if \code{FALSE}, each cohort is reported independently.}

\item{nameStyle}{Name style for the indications. By default:
'indication_\{window_name\}' (mutuallyExclusive = TRUE),
'indication_\{window_name\}_\{cohort_name\}' (mutuallyExclusive = FALSE).}

\item{name}{Name of the new computed cohort table, if NULL a temporary table
will be created.}

\item{restrictIncident}{Whether to include only target cohort records that
start during the analysis period. If FALSE, all target cohort records that
overlap with the study period will be included.}
}
\value{
The original table with a variable added that summarises the
individual´s indications.
}
\description{
Add a variable to a drug cohort indicating their presence in an indication
cohort in a specified time window. If an individual is not in one of the
indication cohorts, they will be considered to have an unknown indication if
they are present in one of the specified OMOP CDM clinical tables. If they
are neither in an indication cohort or a clinical table they will be
considered as having no observed indication.
}
\examples{
\donttest{
library(DrugUtilisation)
library(dplyr, warn.conflicts = FALSE)
library(CDMConnector)

cdm <- mockDrugUtilisation(source = "duckdb")

indications <- list(headache = 378253, asthma = 317009)
cdm <- generateConceptCohortSet(cdm = cdm,
                                conceptSet = indications,
                                name = "indication_cohorts")

cdm <- generateIngredientCohortSet(cdm = cdm,
                                   name = "drug_cohort",
                                   ingredient = "acetaminophen")

cdm$drug_cohort |>
  addIndication(
    indicationCohortName = "indication_cohorts",
    indicationWindow = list(c(0, 0)),
    unknownIndicationTable = "condition_occurrence"
  ) |>
  glimpse()
}

}
