% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{addRoute}
\alias{addRoute}
\title{add route column to a table containing drug_exposure information}
\usage{
addRoute(drugTable, cdm = attr(drugTable, "cdm_reference"))
}
\arguments{
\item{drugTable}{Table in the cdm that must contain drug_concept_id}

\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain 'concept_relationship' table.}
}
\value{
It adds route to the current table
}
\description{
add route column to a table containing drug_exposure information
}
\examples{
\donttest{
library(DrugUtilisation)
library(dplyr)

cdm <- mockDrugUtilisation()

cdm[["drug_exposure"]] \%>\%
  addRoute()
}

}
