% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarkDrugUtilisation.R
\name{benchmarkDrugUtilisation}
\alias{benchmarkDrugUtilisation}
\title{Run benchmark of drug utilisation cohort generation}
\usage{
benchmarkDrugUtilisation(
  cdm,
  ingredient = "acetaminophen",
  alternativeIngredient = c("ibuprofen", "aspirin", "diclofenac"),
  indicationCohort = NULL,
  personSample = 1e+05
)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}

\item{ingredient}{Name of ingredient to benchmark.}

\item{alternativeIngredient}{Name of ingredients to use as alternative
treatments.}

\item{indicationCohort}{Name of a cohort in the cdm_reference object to use
as indicatiomn.}

\item{personSample}{Number of individuals to subset the cdm for the
benchmark. If NULL no sample is applied.}
}
\value{
A summarise_result object.
}
\description{
Run benchmark of drug utilisation cohort generation
}
\examples{
\donttest{
library(DrugUtilisation)
library(omock)

cdm <- mockCdmFromDataset(datasetName = "GiBleed", source = "duckdb")

timings <- benchmarkDrugUtilisation(cdm)

timings
}

}
