% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmSample.R
\name{cdmSample}
\alias{cdmSample}
\title{To subset the cdm object to a certain number of individuals.}
\usage{
cdmSample(cdm, numberIndividuals = 5e+05, tablesToSubset = NULL)
}
\arguments{
\item{cdm}{cdm object.}

\item{numberIndividuals}{number of individuals that are going to be selected
from person table.}

\item{tablesToSubset}{set of tables to subset. If NULL all tables in cdm
object with person_id column are going to be subseted.}
}
\description{
To subset the cdm object to a certain number of individuals.
}
