% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdmSample.R
\name{cdmSubset}
\alias{cdmSubset}
\title{To subset the cdm object tables to the individuals present in a certain
cohort.}
\usage{
cdmSubset(cdm, cohortName, cohortId = NULL, tablesToSubset = NULL)
}
\arguments{
\item{cdm}{cdm object.}

\item{cohortName}{name of the cohort in the cdm reference.}

\item{cohortId}{id of the cohort that we want to subset from the cohort
table. If NULL all cohorts in cohort table are considered.}

\item{tablesToSubset}{set of tables to subset. If NULL all tables in cdm
object with person_id column are going to be subseted.}
}
\description{
To subset the cdm object tables to the individuals present in a certain
cohort.
}
