% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{censorDateDoc}
\alias{censorDateDoc}
\title{Helper for consistent documentation of \code{censorDate}.}
\arguments{
\item{censorDate}{Name of a column that indicates the date to stop the
analysis, if NULL end of individuals observation is used.}
}
\description{
Helper for consistent documentation of \code{censorDate}.
}
\keyword{internal}
