% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cohortDoc}
\alias{cohortDoc}
\title{Helper for consistent documentation of \code{cohort}.}
\arguments{
\item{cohort}{A cohort_table object.}
}
\description{
Helper for consistent documentation of \code{cohort}.
}
\keyword{internal}
