% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cohortIdDoc}
\alias{cohortIdDoc}
\title{Helper for consistent documentation of \code{cohortId}.}
\arguments{
\item{cohortId}{A cohort definition id to restrict by. If NULL, all cohorts
will be included.}
}
\description{
Helper for consistent documentation of \code{cohortId}.
}
\keyword{internal}
