% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Patterns.R
\name{comparePatternsTable}
\alias{comparePatternsTable}
\title{Function to compare a given tibble of drug strength patterns with the
"current one" used in the DrugUtilisation package}
\usage{
comparePatternsTable(pattern_tibble, addId = TRUE)
}
\arguments{
\item{pattern_tibble}{'tibble' containing 'drug_strength' pattern information
to compare with current data in DrugUtilisation package.}

\item{addId}{whether to add pattern ids from tibble in DrugUtilisation
to the given tibble}
}
\value{
List with three arguments:
"diff_patterns" Tibble of differences in patterns between the two tibbles
"diff_numbers" Tibble of differences in numbers of concepts and ingredients
between the two tibbles
"pattern_tibble" If asked, also given tibble with the added column of pattern
numbers used in DrugUtilisation
}
\description{
Function to compare a given tibble of drug strength patterns with the
"current one" used in the DrugUtilisation package
}
