% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{compNameDoc}
\alias{compNameDoc}
\title{Helper for consistent documentation of \code{name} for computed tables.}
\arguments{
\item{name}{Name of the new computed cohort table, if NULL a temporary table
will be created.}
}
\description{
Helper for consistent documentation of \code{name} for computed tables.
}
\keyword{internal}
