% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAggregation.R
\name{computeAggregation}
\alias{computeAggregation}
\title{Explain function}
\usage{
computeAggregation(
  cdm,
  personSummaryName,
  genderAggregation,
  ageGroupsAgregation,
  ageGroups,
  indexYearAggregation,
  indexYearMonthAggregation,
  initialDoseAggregation,
  meanDoseAggregation,
  indicationAggregation,
  indicationTableName,
  aggregationTableName,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{personSummaryName}{personSummaryName}

\item{genderAggregation}{genderAggregation}

\item{ageGroupsAgregation}{ageGroupsAgregation}

\item{ageGroups}{ageGroups}

\item{indexYearAggregation}{indexYearAggregation}

\item{indexYearMonthAggregation}{indexYearMonthAggregation}

\item{initialDoseAggregation}{initialDoseAggregation}

\item{meanDoseAggregation}{meanDoseAggregation}

\item{indicationAggregation}{indicationAggregation}

\item{indicationTableName}{indicationTableName}

\item{aggregationTableName}{aggregationTableName}

\item{verbose}{verbos}
}
\description{
Explain function
}
