% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeIncidencePrevalenceCohorts.R
\name{computeIncidencePrevalenceCohorts}
\alias{computeIncidencePrevalenceCohorts}
\title{Get the cohorts to compute incidence and prevalence}
\usage{
computeIncidencePrevalenceCohorts(
  cdm,
  conceptIds,
  gapEra,
  incidencePrevalenceCohortName,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{gapEra}{gapEra}

\item{incidencePrevalenceCohortName}{incidencePrevalenceCohortName}

\item{verbose}{verbose}

\item{specifications}{specifications}
}
\description{
Get the cohorts to compute incidence and prevalence
}
