% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{daysPrescribedDoc}
\alias{daysPrescribedDoc}
\title{Helper for consistent documentation of \code{daysPrescribed}.}
\arguments{
\item{daysPrescribed}{Whether to include 'days_prescribed' (number of days
prescribed used to create each era).}
}
\description{
Helper for consistent documentation of \code{daysPrescribed}.
}
\keyword{internal}
