% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{defaultTableOptions}
\alias{defaultTableOptions}
\title{Additional arguments for the table functions.}
\usage{
defaultTableOptions()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in table functions,
and their given default values.
}
\examples{
{
  defaultTableOptions()
}

}
