% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{drugRestartFollowUpDaysDoc}
\alias{drugRestartFollowUpDaysDoc}
\title{Helper for consistent documentation of drug-restart \code{followUpDays}.}
\arguments{
\item{followUpDays}{Number of days to follow up. Multiple values are allowed.}
}
\description{
Helper for consistent documentation of drug-restart \code{followUpDays}.
}
\keyword{internal}
