% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{gapEraDoc}
\alias{gapEraDoc}
\title{Helper for consistent documentation of \code{gapEra}.}
\arguments{
\item{gapEra}{Number of days between two continuous exposures to be
considered in the same era.}
}
\description{
Helper for consistent documentation of \code{gapEra}.
}
\keyword{internal}
