% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateCharacterizationTable.R
\name{generateCharacterizationTable}
\alias{generateCharacterizationTable}
\title{Explain function}
\usage{
generateCharacterizationTable(
  cdm,
  tableToCharacterizeName,
  characterizationTableName,
  tableResultsName,
  tableAggregationName,
  minimumCounts,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{tableToCharacterizeName}{tableToCharacterizeName}

\item{characterizationTableName}{characterizationTableName}

\item{tableResultsName}{tableResultsName}

\item{tableAggregationName}{tableAggregationName}

\item{minimumCounts}{minimumCounts}

\item{verbose}{verbose}
}
\description{
Explain function
}
