% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDrugUtilisationCohort.R
\name{generateDrugUtilisationCohort}
\alias{generateDrugUtilisationCohort}
\title{It instantiates the cohorts and their supplementary information
(cohorts_info) for the DUS study}
\usage{
generateDrugUtilisationCohort(
  cdm,
  ingredientConceptId = NULL,
  conceptSetPath = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  summariseMode = "AllEras",
  fixedTime = 365,
  daysPriorHistory = 0,
  gapEra = 30,
  priorUseWashout = NULL,
  imputeDuration = "eliminate",
  durationRange = c(1, NA)
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain at least 'drug_exposure', 'drug_strength' and
observation_period' tables. The 'cdm' object must contain the
'write_schema' as attribute and  the user should have permission to write on
it. It is a compulsory input, no default value is provided.}

\item{ingredientConceptId}{Ingredient OMOP concept that we are interested for
the study. It is a compulsory input, no default value is provided.}

\item{conceptSetPath}{Path to a folder with the concept sets of interest.
Concept sets must be stored in OMOP .json files. If NULL all the descendants
of ingredient concept id will be used. By default: NULL.}

\item{studyStartDate}{Minimum date where the incident exposed eras should
start to be considered. Only incident exposed eras larger than StudyStartDate
are allowed. If it is NULL no restriction is applied. By default: NULL.}

\item{studyEndDate}{Maximum date where the incident exposed eras should
start to be considered. Only incident exposed eras before StudyEndDate
are allowed. If it is NULL no restriction is applied. By default: NULL.}

\item{summariseMode}{Choice on how to summarize the exposures. There are
three options:
"FixedTime" each individual is followed the exact same number of days
specified in 'fixedTime' argument.
"AllEras" we summarize the output will be a summary of the exposed eras of
each individual. Each individual can contribute multiple times.
"FirstEra" we only consider the first observable era of each individual. In
this case each individual can not contribute with multiple rows.
By default: "AllEras".}

\item{fixedTime}{Time period after first exposure where we summarize the
ingredient of interest. Argument only considered if 'summariseMode' =
"FixedTime". No default value is provided.}

\item{daysPriorHistory}{Minimum number of days of prior history
(observation time) required for the incident eras to be considered. By
default: 0, meaning it has to be in observation_period table.
When Null, we do not check if in observation_period table.}

\item{gapEra}{Number of days between two continuous exposures to be
considered in the same era. By default: 180.}

\item{priorUseWashout}{Prior days without exposure. By default: NULL.}

\item{imputeDuration}{Whether/how the duration should be imputed
"eliminate", "median", "mean", "quantile25", "quantile75".
. By default: eliminate}

\item{durationRange}{Range between the duration must be comprised. It should
be a numeric vector of length two, with no NAs and the first value should be
equal or smaller than the second one. It is only required if imputeDuration
= TRUE. If NULL no restrictions are applied. By default: NULL.}
}
\value{
The function returns the 'cdm' object with the created tables as
references of the object.
}
\description{
It instantiates the cohorts and their supplementary information
(cohorts_info) for the DUS study
}
