% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateEpisodeCohortSet.R
\name{generateEpisodeCohortSet}
\alias{generateEpisodeCohortSet}
\title{Generate a cohort set from the CDM episode table}
\usage{
generateEpisodeCohortSet(
  cdm,
  name,
  conceptSet = NULL,
  episodeConceptId = 32531,
  subsetCohort = NULL,
  subsetCohortId = NULL
)
}
\arguments{
\item{cdm}{A \code{cdm_reference} object.}

\item{name}{Name of the new cohort table, it must be a length 1 character
vector.}

\item{conceptSet}{Optional concept set list or numeric vector matching \code{episode_object_concept_id}.
If NULL, all unique \code{episode_object_concept_id} values found in the episode table will be included.}

\item{episodeConceptId}{Numeric vector of concept IDs matching \code{episode_concept_id} (default: 32531 for Treatment Regimen).}

\item{subsetCohort}{Cohort table to subset.}

\item{subsetCohortId}{Cohort definition IDs to use from \code{subsetCohort}.}
}
\value{
A CDM reference object with the new episode cohort table added.
}
\description{
Creates a new cohort table in the CDM reference based on records in the
\code{episode} table, filtered by \code{episode_concept_id} and an optional \code{conceptSet}
matching \code{episode_object_concept_id}.
}
\examples{
\donttest{
library(DrugUtilisation)
cdm <- mockDrugUtilisation()

cdm <- generateEpisodeCohortSet(
  cdm = cdm,
  name = "episode_cohort",
  episodeConceptId = 32531
)
}
}
