% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummarizedTable.R
\name{generateSummarizedTable}
\alias{generateSummarizedTable}
\title{Explain function}
\usage{
generateSummarizedTable(
  cdm,
  tableToSummarizeName,
  variables,
  estimates,
  tableResultsName,
  tableAggregationName,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{specifications}{specifications}
}
\description{
Explain function
}
