% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDoseInformation.R
\name{getDoseInformation}
\alias{getDoseInformation}
\title{Explain function}
\usage{
getDoseInformation(
  cdm,
  dusCohortName,
  conceptSetPath = NULL,
  ingredientConceptId,
  gapEra = 30,
  eraJoinMode = "Previous",
  overlapMode = "Previous",
  sameIndexMode = "Sum",
  imputeDuration = "eliminate",
  imputeDailyDose = "eliminate",
  durationRange = c(1, NA),
  dailyDoseRange = c(0, NA)
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain at least 'drug_exposure', 'drug_strength' and
observation_period' tables.}

\item{dusCohortName}{Name of the cohort that we want to obtain the dose
information.}

\item{conceptSetPath}{Path to a folder with the concept sets of interest.
Concept sets must be stored in OMOP .json files. If NULL all the descendants
of ingredient concept id will be used. By default: NULL.}

\item{ingredientConceptId}{Ingredient OMOP concept that we are interested for
the study. It is a compulsory input, no default value is provided.}

\item{gapEra}{Number of days between two continuous exposures to be
considered in the same era. By default: 180.}

\item{eraJoinMode}{How two different continuous exposures are joined in an
era. There are four options:
"Zero" the exposures are joined considering that the period between both
continuous exposures the subject is treated with a daily dose of zero. The
time between both exposures contributes to the total exposed time.
"Join" the exposures are joined considering that the period between both
continuous exposures the subject is treated with a daily dose of zero. The
time between both exposures does not contribute to the total exposed time.
"Previous" the exposures are joined considering that the period between both
continuous exposures the subject is treated with the daily dose of the
previous subexposure. The time between both exposures contributes to the
total exposed time.
"Subsequent" the exposures are joined considering that the period between
both continuous exposures the subject is treated with the daily dose of the
subsequent subexposure. The time between both exposures contributes to the
total exposed time.
By default: "Previous".}

\item{overlapMode}{How the overlapping between two exposures that do not
start on the same day is solved inside a subexposure. There are five possible
options:
"Previous" the considered daily_dose is the one of the earliest exposure.
"Subsequent" the considered daily_dose is the one of the new exposure that
starts in that subexposure.
"Minimum" the considered daily_dose is the minimum of all of the exposures in
the subexposure.
"Maximum" the considered daily_dose is the maximum of all of the exposures in
the subexposure.
"Sum" the considered daily_dose is the sum of all the exposures present in
the subexposure.
By default: "Previous".}

\item{sameIndexMode}{How the overlapping between two exposures that start on
the same day is solved inside a subexposure. There are five possible options:
"Minimum" the considered daily_dose is the minimum of all of the exposures in
the subexposure.
"Maximum" the considered daily_dose is the maximum of all of the exposures in
the subexposure.
"Sum" the considered daily_dose is the sum of all the exposures present in
the subexposure.
By default: "Sum".}

\item{imputeDuration}{Whether/how the duration should be imputed
"eliminate", "median", "mean", "quantile25", "quantile75".
. By default: eliminate}

\item{imputeDailyDose}{Whether/how the daily_dose should be imputed
"eliminate", "median", "mean", "quantile25", "quantile75". By default:
"eliminate"}

\item{durationRange}{Range between the duration must be comprised. It should
be a numeric vector of length two, with no NAs and the first value should be
equal or smaller than the second one. It is only required if imputeDuration
= TRUE. If NULL no restrictions are applied. By default: NULL.}

\item{dailyDoseRange}{Range between the daily_dose must be comprised. It
should be a numeric vector of length two, with no NAs and the first value
should be equal or smaller than the second one. It is only required if
imputeDailyDose = TRUE. If NULL no restrictions are applied. By default:
NULL.

TERMINOLOGY
\itemize{
\item exposure: we refer to exposure to a row in the drug_exposure table of the
cdm object.
\item subexposure: period of time were the number of exposures that a person is
exposed to does not change.
\item continuous exposure: period of time where the individual is exposed with no
interruption. This period can be comprised by multiple subexposures and
exposure.
\item exposed gap: period of time that the individual is not exposed (there are
no exposures in the time period), but its length is smaller or equal than
gapEra parameter value so the individual is considered to be exposed.
\item non exposed gap: period of time that the individual is not exposed (there are
no exposures in the time period) and its length is larger than gapEra
parameter value
\item era: period of time the individual is considered exposed. It is formed by
continuous exposures and exposed gaps.
}}
}
\value{
The function returns the dose information for all the individuals of
dusCohortName.
}
\description{
Explain function
}
