% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndication.R
\name{getIndication}
\alias{getIndication}
\title{Get indication for a target cohort}
\usage{
getIndication(
  cdm,
  targetCohortName,
  targetCohortDefinitionIds,
  indicationCohortName,
  indicationDefinitionSet,
  indicationGap,
  unknownIndicationTables = c("condition_occurrence", "observation")
)
}
\arguments{
\item{cdm}{name of the cdm object}

\item{targetCohortName}{cohort table in the cdm contain list of target cohort for patients}

\item{targetCohortDefinitionIds}{cohort definition ids to include to generate indication}

\item{indicationCohortName}{indication table that contain list of indication}

\item{indicationDefinitionSet}{definition of indication of interest, this function will only get indication for indication contain in this set}

\item{indicationGap}{the maximum gap between the cohort start date and indication start date}

\item{unknownIndicationTables}{tables to get extra indication from the default is to include condition_occurence and obersvation table in the cdm}
}
\description{
Get indication for a target cohort
}
