% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStratification.R
\name{getStratification}
\alias{getStratification}
\title{Explain function}
\usage{
getStratification(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  sex = NULL,
  ageGroup = NULL,
  indexYearGroup = NULL,
  indicationTable = NULL,
  oneStrata = FALSE
)
}
\arguments{
\item{cdm}{A cdm object created with CDMConnector. No default.}

\item{targetCohortName}{A taget cohort instantiated in the database.}

\item{targetCohortId}{The cohort definition id of the cohort of interest. By
default NULL. NULL only valid when targetCohort contains only one cohort.}

\item{sex}{The sex stratification. By default = NULL.}

\item{ageGroup}{The age groups stratification. By default = NULL.}

\item{indexYearGroup}{The index Years stratification. By default = NULL.}

\item{indicationTable}{The indication table that contains the different
indications.}

\item{oneStrata}{Weather we want to stratify one strata (TRUE) each or
combine in multiple stratas using expand_grid (FALSE). By deafault = FALSE.}
}
\value{
Multiple cohorts as a temporal table in the database. The
stratification for each cohort can be see as attribute ("strata").
}
\description{
Explain function
}
