% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{incidentDoc}
\alias{incidentDoc}
\title{Helper for consistent documentation of \code{incident}.}
\arguments{
\item{incident}{Whether switch treatment must be incident, starting after
discontinuation. If \code{FALSE}, it may start before discontinuation and continue
beyond it.}
}
\description{
Helper for consistent documentation of \code{incident}.
}
\keyword{internal}
