% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{indexDateDoc}
\alias{indexDateDoc}
\title{Helper for consistent documentation of \code{indexDate}.}
\arguments{
\item{indexDate}{Name of a column that indicates the date to start the
analysis.}
}
\description{
Helper for consistent documentation of \code{indexDate}.
}
\keyword{internal}
