% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{indicationCohortIdDoc}
\alias{indicationCohortIdDoc}
\title{Helper for consistent documentation of \code{indicationCohortId}.}
\arguments{
\item{indicationCohortId}{Cohort definition IDs of the indications of
interest. If \code{NULL}, all cohorts in \code{indicationCohortName} are included.}
}
\description{
Helper for consistent documentation of \code{indicationCohortId}.
}
\keyword{internal}
