% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{indicationCohortNameDoc}
\alias{indicationCohortNameDoc}
\title{Helper for consistent documentation of \code{indicationCohortName}.}
\arguments{
\item{indicationCohortName}{Name of the cohort table containing potential
indications.}
}
\description{
Helper for consistent documentation of \code{indicationCohortName}.
}
\keyword{internal}
