% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{indicationWindowDoc}
\alias{indicationWindowDoc}
\title{Helper for consistent documentation of \code{indicationWindow}.}
\arguments{
\item{indicationWindow}{Time windows over which to identify indications.}
}
\description{
Helper for consistent documentation of \code{indicationWindow}.
}
\keyword{internal}
