% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ingredientConceptIdDoc}
\alias{ingredientConceptIdDoc}
\title{Helper for consistent documentation of \code{ingredientConceptId}.}
\arguments{
\item{ingredientConceptId}{Ingredient OMOP concept that we are interested for
the study.}
}
\description{
Helper for consistent documentation of \code{ingredientConceptId}.
}
\keyword{internal}
