% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{inObservationDoc}
\alias{inObservationDoc}
\title{Helper for consistent documentation of \code{inObservation}.}
\arguments{
\item{inObservation}{Whether to restrict the analysis to individuals in
observation. If \code{TRUE}, individuals not in observation are excluded. If
\code{FALSE}, they are included as a separate category.}
}
\description{
Helper for consistent documentation of \code{inObservation}.
}
\keyword{internal}
