% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeScaleCharacterization.R
\name{largeScaleCharacterization}
\alias{largeScaleCharacterization}
\title{Explain function}
\usage{
largeScaleCharacterization(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  temporalWindows,
  tablesToCharacterize,
  characterizationTableName = "characterization",
  permanentCharacterizationTable = FALSE,
  verbose
)
}
\arguments{
\item{cdm}{Connection to a database using DBI::dbConnect()}

\item{targetCohortName}{Name of the table in the cdm that contains the
target cohort}

\item{targetCohortId}{Indentifier for the analyzed target
cohorts}

\item{temporalWindows}{Temporal windows that we want to characterize}

\item{tablesToCharacterize}{Name of the tables in the cdm that we want to
summarize}

\item{characterizationTableName}{characterizationTableName}

\item{permanentCharacterizationTable}{permanentCharacterizationTable}

\item{verbose}{verbose}
}
\description{
Explain function
}
