% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDrugUtilisation.R
\name{mockDrugUtilisation}
\alias{mockDrugUtilisation}
\title{It creates a mock database for testing drugutilisation package}
\usage{
mockDrugUtilisation(
  drug_exposure = NULL,
  drug_strength = NULL,
  drug_concept_id_size = 5,
  ingredient_concept_id_size = 1,
  drug_exposure_size = 10,
  patient_size = 1,
  min_drug_exposure_start_date = "2000-01-01",
  max_drug_exposure_start_date = "2020-01-01",
  seed = 1
)
}
\arguments{
\item{drug_exposure}{default null user can define its own table}

\item{drug_strength}{default null user can define its own table}

\item{drug_concept_id_size}{number of unique drug concept id}

\item{ingredient_concept_id_size}{number of unique drug ingredient concept id}

\item{drug_exposure_size}{number of unique drug exposure}

\item{patient_size}{number of unique patient}

\item{min_drug_exposure_start_date}{user define minimum drug exposure start date}

\item{max_drug_exposure_start_date}{user define maximium drug exposure start date}

\item{seed}{seed}
}
\description{
It creates a mock database for testing drugutilisation package
}
