% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{mutuallyExclusiveDoc}
\alias{mutuallyExclusiveDoc}
\title{Helper for consistent documentation of \code{mutuallyExclusive}.}
\arguments{
\item{mutuallyExclusive}{Whether intersections should be mutually exclusive.
If \code{TRUE}, cohort combinations are reported as mutually exclusive categories;
if \code{FALSE}, each cohort is reported independently.}
}
\description{
Helper for consistent documentation of \code{mutuallyExclusive}.
}
\keyword{internal}
