% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameStyleDoc}
\alias{nameStyleDoc}
\title{Helper for consistent documentation of \code{nameStyle}.}
\arguments{
\item{nameStyle}{Character string to specify the nameStyle of the new columns.}
}
\description{
Helper for consistent documentation of \code{nameStyle}.
}
\keyword{internal}
