% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{newNameDoc}
\alias{newNameDoc}
\title{Helper for consistent documentation of \code{name} for new cohorts.}
\arguments{
\item{name}{Name of the new cohort table, it must be a length 1 character
vector.}
}
\description{
Helper for consistent documentation of \code{name} for new cohorts.
}
\keyword{internal}
