% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{numberExposuresDoc}
\alias{numberExposuresDoc}
\title{Helper for consistent documentation of \code{numberExposures}.}
\arguments{
\item{numberExposures}{Whether to include 'number_exposures' (number of drug
exposure records between indexDate and censorDate).}
}
\description{
Helper for consistent documentation of \code{numberExposures}.
}
\keyword{internal}
