% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{optionsTableDrugUtilisation}
\alias{optionsTableDrugUtilisation}
\title{Additional arguments for the functions tableDrugUtilisation}
\usage{
optionsTableDrugUtilisation()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in tableDoseCoverage,
and their given default values.
}
\examples{
{
optionsTableDrugUtilisation()
}

}
