% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{optionsTableIndication}
\alias{optionsTableIndication}
\title{Additional arguments for the functions tableIndication}
\usage{
optionsTableIndication()
}
\value{
The default .options named list.
}
\description{
It provides a list of allowed inputs for .option argument in tableIndication,
and their given default values.
}
\examples{
{
optionsTableIndication()
}

}
