% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patternsWithFormula}
\alias{patternsWithFormula}
\title{Patterns valid to compute daily dose with the associated formula.}
\format{
A data frame with eight variables: \code{pattern_id}, \code{amount},
\code{amount_unit}, \code{numerator}, \code{numerator_unit},
\code{denominator}, \code{denominator_unit}, \code{formula_name} and
\code{formula}.
}
\usage{
patternsWithFormula
}
\description{
Patterns valid to compute daily dose with the associated formula.
}
\keyword{datasets}
