% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{patternTable}
\alias{patternTable}
\title{Function to create a tibble with the patterns from current drug strength table}
\usage{
patternTable(cdm)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain 'drug_strength' and 'concept' tables.}
}
\value{
The function creates a tibble with the different patterns found in
the table, plus a column of potentially valid and invalid combinations.
}
\description{
Function to create a tibble with the patterns from current drug strength table
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

patternTable(cdm)
}

}
