% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProportionOfPatientsCovered.R
\name{plotProportionOfPatientsCovered}
\alias{plotProportionOfPatientsCovered}
\title{Plot proportion of patients covered}
\usage{
plotProportionOfPatientsCovered(
  result,
  ylim = c(0, NA),
  facet = NULL,
  colour = NULL,
  colour_name = NULL
)
}
\arguments{
\item{result}{Output of summariseProportionOfPatientsCovered}

\item{ylim}{Limits for the Y axis}

\item{facet}{Variables to use for facets}

\item{colour}{Variables to use for colours}

\item{colour_name}{Colour legend name}
}
\value{
Plot of proportion Of patients covered over time
}
\description{
Plot proportion of patients covered
}
