% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportGenerator.R
\docType{data}
\name{reportGenerator}
\alias{reportGenerator}
\title{Information that connects DrugUtilisation with ReportGenerator}
\format{
List of exported elements
}
\usage{
reportGenerator
}
\description{
Information that connects DrugUtilisation with ReportGenerator
}
\examples{
  reportGenerator
}
\keyword{datasets}
