% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{requireDrugInDateRange}
\alias{requireDrugInDateRange}
\title{Restrict to observations that have an indexDate within a certain range.}
\usage{
requireDrugInDateRange(
  cohort,
  dateRange,
  indexDate = "cohort_start_date",
  cohortId = NULL,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{dateRange}{Date interval to consider.}

\item{indexDate}{Column that points to a date column in cohort.}

\item{cohortId}{IDs of the cohorts to modify. The default is NULL meaning all
cohorts will be used; otherwise, only the specified cohorts will be modified,
and the rest will remain unchanged.}

\item{name}{Name of the new cohort. Default name is the original cohort name.}
}
\value{
The cohort table object with the requirements applied.
}
\description{
Restrict to observations that have an indexDate within a certain range.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()
cdm$cohort1 <- cdm$cohort1 |>
  requireDrugInDateRange(
    dateRange = as.Date(c("2020-01-01", NA)), cohortId = 1)
attrition(cdm$cohort1)
}

}
