% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{requireIsFirstDrugEntry}
\alias{requireIsFirstDrugEntry}
\title{Require only first record per subject.}
\usage{
requireIsFirstDrugEntry(
  cohort,
  cohortId = NULL,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{cohortId}{IDs of the cohorts to modify. The default is NULL meaning all
cohorts will be used; otherwise, only the specified cohorts will be modified,
and the rest will remain unchanged.}

\item{name}{Name of the new cohort. Default name is the original cohort name.}
}
\value{
The cohort table object with the requirements applied.
}
\description{
Require only first record per subject.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()
cdm$cohort1 <- cdm$cohort1 |>
  requireIsFirstDrugEntry()
attrition(cdm$cohort1)
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
