% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{requireObservationBeforeDrug}
\alias{requireObservationBeforeDrug}
\title{Require number of days of observation before the drug starts.}
\usage{
requireObservationBeforeDrug(
  cohort,
  priorObservation,
  cohortId = NULL,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{priorObservation}{Number of days of prior observation so the records are
considered.}

\item{cohortId}{IDs of the cohorts to modify. The default is NULL meaning all
cohorts will be used; otherwise, only the specified cohorts will be modified,
and the rest will remain unchanged.}

\item{name}{Name of the new cohort. Default name is the original cohort name.}
}
\value{
The cohort table object with the requirements applied.
}
\description{
Require number of days of observation before the drug starts.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()
cdm$cohort1 <- cdm$cohort1 |>
  requireObservationBeforeDrug(priorObservation = 365, cohortId = 1)
attrition(cdm$cohort1)
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
