% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{restrictIncidentCohortDoc}
\alias{restrictIncidentCohortDoc}
\title{Helper for consistent documentation of restrictIncident in cohort summaries.}
\arguments{
\item{restrictIncident}{Whether to include only target cohort records that
start during the analysis period. If FALSE, all target cohort records that
overlap with the study period will be included.}
}
\description{
Helper for consistent documentation of restrictIncident in cohort summaries.
}
\keyword{internal}
