% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{restrictIncidentDoc}
\alias{restrictIncidentDoc}
\title{Helper for consistent documentation of \code{restrictIncident}.}
\arguments{
\item{restrictIncident}{Whether to include only incident prescriptions in the
analysis. If FALSE all prescriptions that overlap with the study period will
be included.}
}
\description{
Helper for consistent documentation of \code{restrictIncident}.
}
\keyword{internal}
