% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{restrictToFirstDiscontinuationDoc}
\alias{restrictToFirstDiscontinuationDoc}
\title{Helper for consistent documentation of \code{restrictToFirstDiscontinuation}.}
\arguments{
\item{restrictToFirstDiscontinuation}{Whether to consider only the first
discontinuation episode or all of them.}
}
\description{
Helper for consistent documentation of \code{restrictToFirstDiscontinuation}.
}
\keyword{internal}
