% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{ribbonDoc}
\alias{ribbonDoc}
\title{Helper for consistent documentation of \code{ribbon}.}
\arguments{
\item{ribbon}{Whether to plot a ribbon with the confidence intervals.}
}
\description{
Helper for consistent documentation of \code{ribbon}.
}
\keyword{internal}
