% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{seeFormulas}
\alias{seeFormulas}
\title{Formulas applied to compute daily dose.}
\usage{
seeFormulas()
}
\description{
Formulas applied to compute daily dose.
}
\examples{
\donttest{
library(DrugUtilisation)
seeFormulas()
}

}
