% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlUtilities.R
\name{sqlDiffDays}
\alias{sqlDiffDays}
\title{Compute the difference in days between 2 variables in a database table.}
\usage{
sqlDiffDays(dialect, variable1, variable2)
}
\arguments{
\item{dialect}{the database dialect}

\item{variable1}{first date variable}

\item{variable2}{second date variable}
}
\value{
the sql to calculate the difference in days between the 2 variables
}
\description{
Compute the difference in days between 2 variables in a database table.
}
\examples{
\dontrun{
date_diff_sql_code <- sqlDiffDays(dialect = "duckdb",
                                  variable1 = "date_1",
                                  variable2 = "date_2")
}
}
