% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqlUtilities.R
\name{sqlGetAge}
\alias{sqlGetAge}
\title{Calculate accurate age in different sql}
\usage{
sqlGetAge(dialect, dob, dateOfInterest)
}
\arguments{
\item{dialect}{sql dialects: 'Microsoft Sql Server', 'Oracle',
'PostgreSql', 'Amazon RedShift', 'Apache Impala',
'IBM Netezza', 'Google BigQuery', 'Microsoft PDW', 'Snowflake',}

\item{dob}{the name of date of birth column in data}

\item{dateOfInterest}{the name of target date in data}
}
\value{
the sql to calculate age
}
\description{
Calculate accurate age in different sql
}
\examples{
\dontrun{
sql_age <- sqlGetAge(dialect = "duckdb",
                     dob = "dob",
                     date_of_interest = "date")
}
}
