% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{strataDoc}
\alias{strataDoc}
\title{Helper for consistent documentation of \code{strata}.}
\arguments{
\item{strata}{A list of variables to stratify results. These variables
must have been added as additional columns in the cohort table.}
}
\description{
Helper for consistent documentation of \code{strata}.
}
\keyword{internal}
