% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{subsetCohortDoc}
\alias{subsetCohortDoc}
\title{Helper for consistent documentation of \code{subsetCohort}.}
\arguments{
\item{subsetCohort}{Cohort table to subset.}
}
\description{
Helper for consistent documentation of \code{subsetCohort}.
}
\keyword{internal}
