% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{subsetCohortIdDoc}
\alias{subsetCohortIdDoc}
\title{Helper for consistent documentation of \code{subsetCohortId}.}
\arguments{
\item{subsetCohortId}{Cohort definition IDs to use from \code{subsetCohort}.}
}
\description{
Helper for consistent documentation of \code{subsetCohortId}.
}
\keyword{internal}
