% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseDiscontinuationAsSurvival.R
\name{summariseDiscontinuationAsSurvival}
\alias{summariseDiscontinuationAsSurvival}
\title{Summarise discontinuation as a survival analysis}
\usage{
summariseDiscontinuationAsSurvival(
  cohort,
  cohortId = NULL,
  followUpDays = Inf,
  censorDate = NULL,
  restrictToFirstDiscontinuation = TRUE,
  strata = list(),
  competingOutcomeCohortTable = NULL,
  competingOutcomeCohortId = NULL,
  eventGap = 30,
  estimateGap = 1
)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{cohortId}{A cohort definition id to restrict by. If NULL, all cohorts
will be included.}

\item{followUpDays}{Number of days to follow up individuals (lower bound 1,
upper bound Inf).}

\item{censorDate}{if not NULL, an individual's follow up will be censored at
the given date.}

\item{restrictToFirstDiscontinuation}{Whether to consider only the first
discontinuation episode or all of them.}

\item{strata}{A list of variables to stratify results. These variables
must have been added as additional columns in the cohort table.}

\item{competingOutcomeCohortTable}{The competing outcome cohort table of
interest.}

\item{competingOutcomeCohortId}{Competing outcome cohorts to include. It can
either be a cohort_definition_id value or a cohort_name. Multiple ids are
allowed.}

\item{eventGap}{Days between time points for which to report survival events,
which are grouped into the specified intervals.}

\item{estimateGap}{Days between time points for which to report survival
estimates. First day will be day zero with risk estimates provided for times
up to the end of follow-up, with a gap in days equivalent to eventGap.}
}
\value{
A \verb{<summarised_result>} object that contains the probability to not
discontinue over time and the summary statistics. Use
\code{tableDiscontinuationAsSurvival()} and \code{plotDiscontinuationAsSurvival()} to visualise the
results.
}
\description{
\code{summariseDiscontinuationAsSurvival()} analyses discontinuation as a survival analysis
using the \href{https://darwin-eu.github.io/CohortSurvival/}{CohortSurvival}
package. The function assumes that each cohort entry is a continuous
treatment era. Discontinuation will be assessed as a survival analysis with
index date: \emph{start of the drug treatment era} (\code{cohort_start_date}) and event
of interest: \emph{first day without exposure} (the day after
\code{cohort_end_date}). An event is only recorded if the individual remains under
observation on that day. The analysis will use
\code{estimateSingleEventSurvival()} or \code{estimateCompetingRiskSurvival()} depending
if \code{competingOutcomeCohortTable} is provided or not.
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

result <- summariseDiscontinuationAsSurvival(cdm$cohort1)

plotDiscontinuationAsSurvival(result)

tableDiscontinuationAsSurvival(result)
}

}
