% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseLargeScaleCharacteristics.R
\name{summariseLargeScaleCharacteristics}
\alias{summariseLargeScaleCharacteristics}
\title{This function has been deprecated please use CohortCharacteristics package}
\usage{
summariseLargeScaleCharacteristics()
}
\description{
This function has been deprecated please use CohortCharacteristics package
}
