% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseProportionOfPatientsCovered.R
\name{summariseProportionOfPatientsCovered}
\alias{summariseProportionOfPatientsCovered}
\title{summarise proportion Of patients covered}
\usage{
summariseProportionOfPatientsCovered(
  cohort,
  cohortId = NULL,
  strata = list(),
  followUpDays = NULL
)
}
\arguments{
\item{cohort}{A cohort table}

\item{cohortId}{Cohort definition id of interest. If NUll all cohorts will
be considered.}

\item{followUpDays}{Number of days to follow up individuals for. If NULL the
maximum amount of days from an individuals first cohort start date to their
last cohort end date will be used}
}
\value{
A summarised result
}
\description{
summarise proportion Of patients covered
}
