% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseT1.R
\name{summariseT1}
\alias{summariseT1}
\title{It instantiates the cohorts and their supplementary information
(cohorts_info) for the DUS study}
\usage{
summariseT1(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  covariatesTableName = NULL,
  covariatesSet = NULL,
  covariatesWindow = NULL,
  ...
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain at least 'drug_exposure', 'drug_strength' and
observation_period' tables. The 'cdm' object must contain the
'write_schema' as attribute and  the user should have permission to write on
it. It is a compulsory input, no default value is provided.}

\item{targetCohortName}{Name of the target cohort}

\item{targetCohortId}{id or vector of ids of the cohort definitions in
targetCohort that we are interested in. If NULL all cohort_definitions from
targetCohort are considered By default: NULL.}

\item{covariatesTableName}{covariatesTableName}

\item{covariatesSet}{covariatesSet}

\item{covariatesWindow}{covariatesWindow}

\item{...}{add as many covariates groups as you want}
}
\value{
The function returns the 'cdm' object with the created tables as
references of the object.
}
\description{
It instantiates the cohorts and their supplementary information
(cohorts_info) for the DUS study
}
