% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTreatmentDiscontinuation.R
\name{summariseTreatmentDiscontinuation}
\alias{summariseTreatmentDiscontinuation}
\title{Summarise treatment discontinuation}
\usage{
summariseTreatmentDiscontinuation(
  cdm,
  targetCohortTable,
  targetCohortId = NULL,
  outcomeWashout = Inf,
  censorOnDate = NULL,
  followUpDays = Inf,
  strata = NULL,
  eventGap = 30,
  estimateGap = 1,
  restrictedMeanFollowUp = NULL,
  minimumSurvivalDays = 1,
  minCellCount = 5
)
}
\arguments{
\item{cdm}{CDM reference}

\item{targetCohortTable}{targetCohortTable}

\item{targetCohortId}{targetCohortId}

\item{outcomeWashout}{Washout time in days for the outcome}

\item{censorOnDate}{if not NULL, an individual's follow up will be censored
at the given date}

\item{followUpDays}{Number of days to follow up individuals (lower bound 1,
upper bound Inf)}

\item{strata}{strata}

\item{eventGap}{Days between time points for which to report survival
events, which are grouped into the specified intervals.}

\item{estimateGap}{Days between time points for which to report survival
estimates. First day will be day zero with risk estimates provided
for times up to the end of follow-up, with a gap in days equivalent
to eventGap.}

\item{restrictedMeanFollowUp}{number of days of follow-up to take into account
when calculating restricted mean for all cohorts}

\item{minimumSurvivalDays}{Minimum number of days required for the main cohort
to have survived}

\item{minCellCount}{The minimum number of events to reported, below which
results will be obscured. If 0, all results will be reported.}

\item{censorOnCohortExit}{If TRUE, an individual's follow up will be
censored at their cohort exit}
}
\value{
tibble with discontinuation information for the drug cohort,
including: time, people at risk, survival probability, cumulative incidence,
95 CIs, strata and outcome. A tibble with the number of events is
outputted as an attribute of the output
}
\description{
Summarise treatment discontinuation
}
