% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTreatment.R
\name{summariseTreatmentFromConceptSet}
\alias{summariseTreatmentFromConceptSet}
\title{This function is used to summarise the dose table over multiple cohorts.}
\usage{
summariseTreatmentFromConceptSet(
  cohort,
  strata = list(),
  window,
  treatmentConceptSet = NULL,
  combination = FALSE,
  minCellCount = 5
)
}
\arguments{
\item{cohort}{Cohort with drug use variables and strata.}

\item{strata}{Stratification list.}

\item{window}{Window where to summarise the treatments.}

\item{treatmentConceptSet}{Concept set list to summarise.}

\item{combination}{Whether to include combination treatments.}

\item{minCellCount}{Below this number counts will be suppressed.}
}
\value{
A summary of the drug use stratified by cohort_name and strata_name
}
\description{
This function is used to summarise the dose table over multiple cohorts.
}
