% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{switchCohortIdDoc}
\alias{switchCohortIdDoc}
\title{Helper for consistent documentation of \code{switchCohortId}.}
\arguments{
\item{switchCohortId}{Cohort definition IDs to use from \code{switchCohortTable}.
If \code{NULL}, all cohort definition IDs are included.}
}
\description{
Helper for consistent documentation of \code{switchCohortId}.
}
\keyword{internal}
