% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{switchCohortTableDoc}
\alias{switchCohortTableDoc}
\title{Helper for consistent documentation of \code{switchCohortTable}.}
\arguments{
\item{switchCohortTable}{A cohort table in the CDM containing possible
alternative treatments.}
}
\description{
Helper for consistent documentation of \code{switchCohortTable}.
}
\keyword{internal}
