% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{treatmentCohortIdDoc}
\alias{treatmentCohortIdDoc}
\title{Helper for consistent documentation of \code{treatmentCohortId}.}
\arguments{
\item{treatmentCohortId}{Cohort definition IDs of the treatments of interest.
If \code{NULL}, all cohorts in \code{treatmentCohortName} are included.}
}
\description{
Helper for consistent documentation of \code{treatmentCohortId}.
}
\keyword{internal}
