% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{treatmentCohortNameDoc}
\alias{treatmentCohortNameDoc}
\title{Helper for consistent documentation of \code{treatmentCohortName}.}
\arguments{
\item{treatmentCohortName}{Name of the cohort table containing the treatments
of interest.}
}
\description{
Helper for consistent documentation of \code{treatmentCohortName}.
}
\keyword{internal}
