% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentDiscontinuation.R
\name{treatmentDiscontinuation}
\alias{treatmentDiscontinuation}
\title{Compute the treatment dicontinuation.}
\usage{
treatmentDiscontinuation(
  cohort,
  cohortId = NULL,
  strata = list(),
  censorOnDate = NULL,
  followUpDays = Inf,
  timeGap = c(1, 7, 30, 365),
  times = NULL,
  minCellCount = 5
)
}
\arguments{
\item{cohort}{GeneratedCohortSet object.}

\item{cohortId}{cohort_definition_id to target.}

\item{strata}{List of columns or combinations of columns to stratify for.}

\item{censorOnDate}{##COPY FROM CS##}

\item{timeGap}{##COPY FROM CS##}

\item{times}{##COPY FROM CS##}

\item{minCellCount}{##COPY FROM CS##}

\item{folloUpDays}{##COPY FROM CS##}
}
\description{
Compute the treatment dicontinuation.
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

cdm <- generateDrugUtilisationCohortSet(
  cdm = cdm, name = "cohort", conceptSet = list("acetaminophen" = 1214)
)

treatmentDiscontinuation(cdm$cohort)
}

}
