% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{unknownIndicationTableDoc}
\alias{unknownIndicationTableDoc}
\title{Helper for consistent documentation of \code{unknownIndicationTable}.}
\arguments{
\item{unknownIndicationTable}{Tables in the OMOP CDM to search for unknown
indications.}
}
\description{
Helper for consistent documentation of \code{unknownIndicationTable}.
}
\keyword{internal}
