% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{windowDoc}
\alias{windowDoc}
\title{Helper for consistent documentation of \code{window}.}
\arguments{
\item{window}{Time windows over which to identify treatments.}
}
\description{
Helper for consistent documentation of \code{window}.
}
\keyword{internal}
