% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDailyDose.R
\name{addDailyDose}
\alias{addDailyDose}
\title{Explain function}
\usage{
addDailyDose(table, cdm, ingredientConceptId, tablePrefix = NULL)
}
\arguments{
\item{table}{table}

\item{cdm}{cdm}

\item{ingredientConceptId}{ingredientConceptId}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\description{
Explain function
}
