% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDailyDose.R
\name{addDailyDose}
\alias{addDailyDose}
\title{Explain function}
\usage{
addDailyDose(table, cdm, ingredientConceptId)
}
\arguments{
\item{table}{table}

\item{cdm}{cdm}

\item{ingredientConceptId}{ingredientConceptId}
}
\description{
Explain function
}
