% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyDose.R
\name{addDailyDose}
\alias{addDailyDose}
\title{Add daily dose to \code{drug_exposure} like table}
\usage{
addDailyDose(x, ingredientConceptId, name = NULL)
}
\arguments{
\item{x}{A \code{cdm_table} class table with at least \code{drug_concept_id},
\code{drug_exposure_start_date}, \code{drug_exposure_end_date} and \code{quantity} as
columns.}

\item{ingredientConceptId}{Ingredient OMOP concept that we are interested for
the study.}

\item{name}{Name of the new generated table, if \code{NULL} a temporary table will
be generated.}
}
\value{
A \code{cdm_table} with two new columns 'daily_dose' and 'unit'.
}
\description{
Add daily dose to \code{drug_exposure} like table
}
