% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIndication.R
\name{addIndication}
\alias{addIndication}
\title{Get indication for a target cohort}
\usage{
addIndication(
  x,
  cdm,
  indicationCohortName,
  indicationGap = 0,
  unknownIndicationTable = c("condition_occurrence", "observation"),
  indicationDate = "cohort_start_date"
)
}
\arguments{
\item{x}{Table in the cdm}

\item{cdm}{A cdm reference created using CDMConnector}

\item{indicationCohortName}{Name of indication cohort table}

\item{indicationGap}{Gap between the event and the indication}

\item{unknownIndicationTable}{Tables to search unknown indications}

\item{indicationDate}{Date of the indication}
}
\value{
Same cohort adding the indications
}
\description{
Get indication for a target cohort
}
\examples{
\donttest{
library(DrugUtilisation)
library(dplyr)

cdm <- mockDrugUtilisation()
cdm$cohort1 \%>\%
  addIndication(
    cdm = cdm, indicationCohortName = "cohort2", indicationGap = c(0, 30)
  ) \%>\%
  collect()
}

}
