% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIndication.R
\name{addIndication}
\alias{addIndication}
\title{Get indication for a target cohort}
\usage{
addIndication(
  x,
  indicationCohortName,
  indicationCohortId = NULL,
  indicationWindow = list(c(0, 0)),
  unknownIndicationTable = NULL,
  indexDate = "cohort_start_date",
  name = NULL
)
}
\arguments{
\item{x}{Table in the cdm}

\item{indicationCohortName}{Name of indication cohort table}

\item{indicationCohortId}{target cohort Id to add indication}

\item{indicationWindow}{time window of interests}

\item{unknownIndicationTable}{Tables to search unknown indications}

\item{indexDate}{Date of the indication}

\item{name}{name of permenant table}
}
\value{
Same cohort adding the indications
}
\description{
Get indication for a target cohort
}
\examples{
\donttest{
library(DrugUtilisation)
library(CDMConnector)

cdm <- mockDrugUtilisation()

indications <- list("headache" = 378253, "asthma" = 317009)
cdm <- generateConceptCohortSet(
  cdm, indications, "indication_cohorts"
)

acetaminophen <- getDrugIngredientCodes(cdm, "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(cdm, "drug_cohort", acetaminophen)

cdm[["drug_cohort"]] |>
  addIndication("indication_cohorts", indicationWindow = list(c(0,0)))
}

}
