% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{addIndication}
\alias{addIndication}
\title{Add a variable indicating individuals indications}
\usage{
addIndication(
  cohort,
  indicationCohortName,
  indicationCohortId = NULL,
  indicationWindow = list(c(0, 0)),
  unknownIndicationTable = NULL,
  indexDate = "cohort_start_date",
  censorDate = NULL,
  mutuallyExclusive = TRUE,
  nameStyle = NULL,
  name = NULL
)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{indicationCohortName}{Name of indication cohort table}

\item{indicationCohortId}{target cohort Id to add indication}

\item{indicationWindow}{time window of interests}

\item{unknownIndicationTable}{Tables to search unknown indications}

\item{indexDate}{Name of a column that indicates the date to start the
analysis.}

\item{censorDate}{Name of a column that indicates the date to stop the
analysis, if NULL end of individuals observation is used.}

\item{mutuallyExclusive}{Whether to consider mutually exclusive categories
(one column per window) or not (one column per window and indication).}

\item{nameStyle}{Name style for the indications. By default:
'indication_\{window_name\}' (mutuallyExclusive = TRUE),
'indication_\{window_name\}_\{cohort_name\}' (mutuallyExclusive = FALSE).}

\item{name}{Name of the new computed cohort table, if NULL a temporary table
will be created.}
}
\value{
The original table with a variable added that summarises the
individual´s indications.
}
\description{
Add a variable to a drug cohort indicating their presence in an indication
cohort in a specified time window. If an individual is not in one of the
indication cohorts, they will be considered to have an unknown indication if
they are present in one of the specified OMOP CDM clinical tables. If they
are neither in an indication cohort or a clinical table they will be
considered as having no observed indication.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()

indications <- list("headache" = 378253, "asthma" = 317009)
cdm <- CDMConnector::generateConceptCohortSet(
  cdm = cdm, conceptSet = indications, name = "indication_cohorts"
)

cdm <- generateIngredientCohortSet(
  cdm = cdm, name = "drug_cohort",
  ingredient = "acetaminophen"
)

cdm$drug_cohort |>
  addIndication(
    indicationCohortName = "indication_cohorts",
    indicationWindow = list(c(0, 0)),
    unknownIndicationTable = "condition_occurrence"
  ) |>
  dplyr::glimpse()
}

}
