% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPattern.R
\name{addPattern}
\alias{addPattern}
\title{add pattern info to a table containing drug_strength information}
\usage{
addPattern(table)
}
\arguments{
\item{table}{table}

\item{cdm}{cdm}

\item{ingredientConceptId}{ingredientConceptId}

\item{tablePrefix}{The stem for the permanent tables that will
be created. If NULL, temporary tables will be used throughout.}
}
\description{
add pattern info to a table containing drug_strength information
}
