% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPattern.R
\name{addPattern}
\alias{addPattern}
\title{add pattern info to a table containing drug_strength information}
\usage{
addPattern(table)
}
\arguments{
\item{table}{table}
}
\value{
table with a pattern_id and unit columns
}
\description{
add pattern info to a table containing drug_strength information
}
