% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIntersect.R
\name{addTreatment}
\alias{addTreatment}
\title{Add a variable indicating individuals medications}
\usage{
addTreatment(
  cohort,
  treatmentCohortName,
  treatmentCohortId = NULL,
  window = list(c(0, 0)),
  indexDate = "cohort_start_date",
  censorDate = NULL,
  mutuallyExclusive = TRUE,
  nameStyle = NULL,
  name = NULL,
  restrictIncident = FALSE
)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{treatmentCohortName}{Name of the cohort table containing the treatments
of interest.}

\item{treatmentCohortId}{Cohort definition IDs of the treatments of interest.
If \code{NULL}, all cohorts in \code{treatmentCohortName} are included.}

\item{window}{Time windows over which to identify treatments.}

\item{indexDate}{Name of a column that indicates the date to start the
analysis.}

\item{censorDate}{Name of a column that indicates the date to stop the
analysis, if NULL end of individuals observation is used.}

\item{mutuallyExclusive}{Whether intersections should be mutually exclusive.
If \code{TRUE}, cohort combinations are reported as mutually exclusive categories;
if \code{FALSE}, each cohort is reported independently.}

\item{nameStyle}{Name style for the treatment columns. By default:
'treatment_\{window_name\}' (mutuallyExclusive = TRUE),
'treatment_\{window_name\}_\{cohort_name\}' (mutuallyExclusive = FALSE).}

\item{name}{Name of the new computed cohort table, if NULL a temporary table
will be created.}

\item{restrictIncident}{Whether to include only target cohort records that
start during the analysis period. If \code{FALSE}, all target cohort records
that overlap with the study period will be included.}
}
\value{
The original table with a variable added that summarises the
individual´s indications.
}
\description{
Add a variable to a drug cohort indicating their presence of a medication
cohort in a specified time window.
}
\examples{
\donttest{
library(DrugUtilisation)
library(dplyr, warn.conflicts = FALSE)

cdm <- mockDrugUtilisation(numberIndividuals = 50)

cdm <- generateIngredientCohortSet(cdm = cdm,
                                   name = "drug_cohort",
                                   ingredient = "acetaminophen")

cdm <- generateIngredientCohortSet(cdm = cdm,
                                   name = "treatments",
                                   ingredient = c("metformin", "simvastatin"))

cdm$drug_cohort |>
  addTreatment("treatments", window = list(c(0, 0), c(1, 30), c(31, 60))) |>
  glimpse()
}

}
