% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAggregation.R
\name{computeAggregation}
\alias{computeAggregation}
\title{Explain function}
\usage{
computeAggregation(
  cdm,
  personSummaryName,
  genderAggregation,
  ageGroupsAgregation,
  ageGroups,
  yearAggregation,
  yearMonthAggregation,
  initialDoseAggregation,
  meanDoseAggregation,
  indicationAggregation,
  indicationTableName,
  aggregationTableName,
  permanentAggregationTable,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{specifications}{specifications}
}
\description{
Explain function
}
