% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAggregation.R
\name{computeAggregation}
\alias{computeAggregation}
\title{Explain function}
\usage{
computeAggregation(
  cdm,
  personSummaryName,
  genderAggregation = TRUE,
  ageGroupsAgregation = TRUE,
  ageGroups = list(c(0, 19), c(20, 39), c(40, 59), c(60, 79), c(80, 150)),
  indexYearAggregation = TRUE,
  indexYearMonthAggregation = FALSE,
  initialDoseAggregation = TRUE,
  meanDoseAggregation = FALSE,
  cohortid = NULL,
  aggregationTableName = "drug_utilisation_aggregation_table",
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{personSummaryName}{personSummaryName}

\item{genderAggregation}{genderAggregation}

\item{ageGroupsAgregation}{ageGroupsAgregation}

\item{ageGroups}{ageGroups the input is a list of user define age group}

\item{indexYearAggregation}{indexYearAggregation}

\item{indexYearMonthAggregation}{indexYearMonthAggregation}

\item{initialDoseAggregation}{initialDoseAggregation}

\item{meanDoseAggregation}{meanDoseAggregation}

\item{aggregationTableName}{aggregationTableName}

\item{verbose}{verbose}

\item{indicationAggregation}{indicationAggregation}

\item{indicationTableName}{indicationTableName}
}
\description{
Explain function
}
