% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortUtilities.R
\name{computeCohortAttrition}
\alias{computeCohortAttrition}
\title{Add a line in the attrition table. If the table does not exist it is created}
\usage{
computeCohortAttrition(
  x,
  cdm,
  attrition = NULL,
  reason = "Qualifying initial events"
)
}
\arguments{
\item{x}{A table in the cdm with at lest: 'cohort_definition_id' and
subject_id'}

\item{cdm}{A cdm reference created using CDMConnector}

\item{attrition}{An attrition table. If NULL a new attrition table is created.}

\item{reason}{A character with the name of the reason.}
}
\value{
The function returns the 'cdm' object with the created tables as
references of the object.
}
\description{
Add a line in the attrition table. If the table does not exist it is created
}
