% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortUtilities.R
\name{computeCohortCount}
\alias{computeCohortCount}
\title{Computes the cohortCount attribute for a certain table}
\usage{
computeCohortCount(x, cdm)
}
\arguments{
\item{x}{A table in the cdm with at lest: 'cohort_definition_id' and
subject_id'}

\item{cdm}{A cdm_reference object}

\item{name}{Name of the generated table}
}
\value{
A reference to a table in the database with the cohortCount
}
\description{
Computes the cohortCount attribute for a certain table
}
