% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDailyDose_New.R
\name{computeDailyDose}
\alias{computeDailyDose}
\title{Explain function}
\usage{
computeDailyDose(cdm, tableName, ingredientConceptId)
}
\arguments{
\item{cdm}{cdm}

\item{tableName}{tableName}

\item{ingredientConceptId}{ingredientConceptId}
}
\description{
Explain function
}
