% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDailyDose.R
\name{computeDailyDose}
\alias{computeDailyDose}
\title{Explain function}
\usage{
computeDailyDose(
  table,
  cdm = cdm,
  ingredient_concept_id = ingredient_concept_id,
  verbose = FALSE
)
}
\arguments{
\item{table}{table}

\item{cdm}{cdm}

\item{ingredient_concept_id}{ingredient_concept_id}

\item{verbose}{verbose}
}
\description{
Explain function
}
