% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dailyDose.R
\name{dailyDoseCoverage}
\alias{dailyDoseCoverage}
\title{Check coverage of daily dose computation in a sample of the cdm for selected
concept sets and ingredient}
\usage{
dailyDoseCoverage(cdm, ingredientConceptId)
}
\arguments{
\item{cdm}{A cdm reference created using CDMConnector}

\item{ingredientConceptId}{Code indicating the ingredient of interest}
}
\value{
The function returns information of the coverage of computeDailyDose.R
for the selected ingredients and concept sets
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
