% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseDrugUse.R
\name{drugUseColumns}
\alias{drugUseColumns}
\title{Obtain automatically the drug use columns}
\usage{
drugUseColumns(x)
}
\arguments{
\item{x}{Tibble}
}
\value{
Name of the drug use columns
}
\description{
Obtain automatically the drug use columns
}
\examples{
\donttest{
library(drugUtilisation)
library(CodelistGenerator)

cdm <- mockDrugUtilisation()
cdm <- generateDrugUtilisationCohortSet(
  cdm, "dus", getDrugIngredientCodes(cdm, "acetaminophen")
)
cdm$dus <- cdm$dus \%>\%
  addDrugUse(cdm, 1125315)
drugUseColumns(cdm$dus)
}
}
