% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrugUtilisation-package.R
\docType{package}
\name{DrugUtilisation-package}
\alias{DrugUtilisation}
\alias{DrugUtilisation-package}
\title{DrugUtilisation: Summarise patient-level drug utilisation in data mapped to the OMOP Common Data Model}
\description{
Summarise patient-level drug utilisation in data mapped to the OMOP Common Data Model.
}
\author{
\strong{Maintainer}: Marti Catala \email{marti.catalasabate@ndorms.ox.ac.uk}

}
\keyword{internal}
