% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateIngredientCohortSet.R
\name{generateIngredientCohortSet}
\alias{generateIngredientCohortSet}
\title{Generates a cohort of the drug use of ingredient name(s) of interest.}
\usage{
generateIngredientCohortSet(
  cdm,
  name,
  ingredient = NULL,
  durationRange = lifecycle::deprecated(),
  imputeDuration = lifecycle::deprecated(),
  gapEra = 0,
  priorUseWashout = 0,
  priorObservation = lifecycle::deprecated(),
  cohortDateRange = lifecycle::deprecated(),
  limit = lifecycle::deprecated(),
  doseForm = NULL,
  ingredientRange = c(1, Inf)
)
}
\arguments{
\item{cdm}{A cdm_reference object.}

\item{name}{Name of the GeneratedCohortSet}

\item{ingredient}{Accepts both vectors and named lists of ingredient names.
For a vector input, e.g., c("acetaminophen", "codeine"), it generates a
cohort table with descendant concept codes for each ingredient, assigning
unique cohort_definition_id. For a named list input, e.g., list(
"test_1" = c("simvastatin", "acetaminophen"), "test_2" = "metformin"),
it produces a cohort table based on the structure of the input, where
each name leads to a combined set of descendant concept codes for the
specified ingredients, creating distinct cohort_definition_id for each
named group.}

\item{durationRange}{Deprecated.}

\item{imputeDuration}{Deprecated.}

\item{gapEra}{Number of days between two continuous exposures to be
considered in the same era.}

\item{priorUseWashout}{Prior days without exposure.}

\item{priorObservation}{Deprecated.}

\item{cohortDateRange}{Deprecated.}

\item{limit}{Deprecated.}

\item{doseForm}{Only descendants codes with the specified dose form
will be returned. If NULL, descendant codes will be returned regardless
of dose form.}

\item{ingredientRange}{Used to restrict descendant codes to those
associated with a specific number of ingredients. Must be a vector of length
two with the first element the minimum number of ingredients allowed and
the second the maximum. A value of c(2, 2) would restrict to only concepts
associated with two ingredients.}
}
\value{
The function returns the 'cdm' object with the created cohorts as
references of the object.
}
\description{
Generates a cohort of the drug use of ingredient name(s) of interest.
}
\examples{
\donttest{
library(DrugUtilisation)
cdm <- mockDrugUtilisation()
cdm <- generateIngredientCohortSet(
  cdm = cdm,
  ingredient = "acetaminophen",
  name = "test"
)
cdm
}

}
