% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSummarizedTable.R
\name{generateSummarizedTable}
\alias{generateSummarizedTable}
\title{Explain function}
\usage{
generateSummarizedTable(
  cdm,
  tableToSummarizeName,
  variables,
  estimates,
  tableResultsName,
  tableAggregationName,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{tableToSummarizeName}{tableToSummarizeName}

\item{variables}{variables}

\item{estimates}{estimates}

\item{tableResultsName}{tableResultsName}

\item{tableAggregationName}{tableAggregationName}

\item{verbose}{verbose}
}
\description{
Explain function
}
