% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndication.R
\name{getIndication}
\alias{getIndication}
\title{Get indication for a target cohort}
\usage{
getIndication(
  cdm,
  targetCohortName,
  targetCohortDefinitionIds = NULL,
  indicationCohortName,
  indicationCohortDefinitionIds = NULL,
  indicationDefinitionSet,
  indicationGap = 0,
  unknownIndication = TRUE,
  unknownIndicarionTables = c("condition_occurrence", "observation"),
  indicationTableName = "indication_table",
  verbose = FALSE
)
}
\arguments{
\item{cdm}{cdm}

\item{targetCohortName}{targetCohortName}

\item{targetCohortDefinitionIds}{targetCohortDefinitionIds}

\item{indicationCohortName}{indicationCohortName}

\item{indicationCohortDefinitionIds}{indicationCohortDefinitionIds}

\item{indicationDefinitionSet}{indicationDefinitionSet}

\item{indicationGap}{indicationGap}

\item{unknownIndication}{unknownIndication}

\item{unknownIndicarionTables}{unknownIndicarionTables}

\item{indicationTableName}{indicationTableName}

\item{verbose}{verbose}

\item{noIndication}{noIndication}
}
\description{
Get indication for a target cohort
}
