% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseIndication.R
\name{indicationColumns}
\alias{indicationColumns}
\title{Obtain automatically the indication columns}
\usage{
indicationColumns(x)
}
\arguments{
\item{x}{Tibble}
}
\value{
Name of the indication columns
}
\description{
Obtain automatically the indication columns
}
