% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiateDrugUtilisationCohorts.R
\name{instantiateDrugUtilisationCohorts}
\alias{instantiateDrugUtilisationCohorts}
\title{Explain function}
\usage{
instantiateDrugUtilisationCohorts(
  cdm,
  specifications,
  studyTime = NULL,
  drugUtilisationCohortName,
  imputeDuration = FALSE,
  imputeDailyDose = FALSE,
  durationLowerBound = NULL,
  durationUpperBound = NULL,
  dailyDoseLowerBound = NULL,
  dailyDoseUpperBound = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{Connection to a database using DBI::dbConnect()}

\item{verbose}{verbose}

\item{targetCohortName}{Name of the table in the cdm that contains the
target cohort}

\item{targetCohortId}{Indentifier for the analyzed target
cohorts}

\item{temporalWindows}{Temporal windows that we want to characterize}

\item{tablesToCharacterize}{Name of the tables in the cdm that we want to
summarize}

\item{characterizationTableName}{characterizationTableName}
}
\description{
Explain function
}
