% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiateDrugUtilisationCohorts.R
\name{instantiateDrugUtilisationCohorts}
\alias{instantiateDrugUtilisationCohorts}
\title{Explain function}
\usage{
instantiateDrugUtilisationCohorts(
  cdm,
  specifications,
  studyTime = NULL,
  gapEra,
  eraJoinMode,
  overlapMode,
  sameIndexMode,
  drugUtilisationCohortName,
  imputeDuration = FALSE,
  imputeDailyDose = FALSE,
  durationLowerBound = NULL,
  durationUpperBound = NULL,
  dailyDoseLowerBound = NULL,
  dailyDoseUpperBound = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{cdm}

\item{specifications}{specifications}

\item{studyTime}{studyTime}

\item{gapEra}{gapEra}

\item{eraJoinMode}{eraJoinMode}

\item{overlapMode}{overlapMode}

\item{sameIndexMode}{sameIndexMode}

\item{drugUtilisationCohortName}{drugUtilisationCohortName}

\item{imputeDuration}{imputeDuration}

\item{imputeDailyDose}{imputeDailyDose}

\item{durationLowerBound}{durationLowerBound}

\item{durationUpperBound}{durationUpperBound}

\item{dailyDoseLowerBound}{dailyDoseLowerBound}

\item{dailyDoseUpperBound}{dailyDoseUpperBound}

\item{verbose}{verbose}
}
\description{
Explain function
}
