% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiateDrugUtilisationCohorts.R
\name{instantiateDrugUtilisationCohorts}
\alias{instantiateDrugUtilisationCohorts}
\title{It instantiates the cohorts and their supplementary information
(cohorts_info) for the DUS study}
\usage{
instantiateDrugUtilisationCohorts(
  cdm,
  specifications = NULL,
  ingredientConceptId,
  studyTime = 365,
  gapEra = 30,
  eraJoinMode = "first",
  overlapMode = "first",
  sameIndexMode = "sum",
  drugUtilisationCohortName,
  imputeDuration = FALSE,
  imputeDailyDose = FALSE,
  durationLowerBound = NULL,
  durationUpperBound = NULL,
  dailyDoseLowerBound = NULL,
  dailyDoseUpperBound = NULL,
  cohortEntryPriorHistory = NULL,
  studyStartDate = NULL,
  studyEndDate = NULL,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{cdm}

\item{specifications}{specifications}

\item{ingredientConceptId}{Ingredient OMOP concept that we are interested in the study}

\item{studyTime}{studyTime}

\item{gapEra}{gapEra}

\item{eraJoinMode}{ways to join drug exposures with gap to drug era}

\item{overlapMode}{ways to deal with overlapping exposures}

\item{sameIndexMode}{ways to join exposures when having same start dates}

\item{drugUtilisationCohortName}{drugUtilisationCohortName}

\item{imputeDuration}{imputeDuration}

\item{imputeDailyDose}{imputeDailyDose}

\item{durationLowerBound}{durationLowerBound}

\item{durationUpperBound}{durationUpperBound}

\item{dailyDoseLowerBound}{dailyDoseLowerBound}

\item{dailyDoseUpperBound}{dailyDoseUpperBound}

\item{cohortEntryPriorHistory}{minimum prior history/observation time}

\item{studyStartDate}{cohort level criteria, cohort start date in result: larger than StudyStartDate}

\item{studyEndDate}{cohort level criteria, cohort start date in result: smaller than StudyStartDate}

\item{verbose}{verbose}
}
\description{
It instantiates the cohorts and their supplementary information
(cohorts_info) for the DUS study
}
