% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiateIncidencePrevalenceCohorts.R
\name{instantiateIncidencePrevalenceCohorts}
\alias{instantiateIncidencePrevalenceCohorts}
\title{Get the cohorts to compute incidence and prevalence}
\usage{
instantiateIncidencePrevalenceCohorts(
  cdm,
  conceptIds,
  gapEra,
  incidencePrevalenceCohortName,
  cohortDefinitionId = NULL,
  append = FALSE,
  overWrite = TRUE,
  verbose
)
}
\arguments{
\item{cdm}{drug_exposure table from cdm}

\item{conceptIds}{conceptIds users want to compute era}

\item{gapEra}{Maxmium gap for the gap Era}

\item{incidencePrevalenceCohortName}{incidencePrevalenceCohortName}

\item{cohortDefinitionId}{cohortDefinitionId}

\item{append}{append or compute table in cdm}

\item{overWrite}{overWrite}

\item{verbose}{verbose}
}
\description{
Get the cohorts to compute incidence and prevalence
}
