% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instantiateIncidencePrevalenceCohorts.R
\name{instantiateIncidencePrevalenceCohorts}
\alias{instantiateIncidencePrevalenceCohorts}
\title{Get the cohorts to compute incidence and prevalence}
\usage{
instantiateIncidencePrevalenceCohorts(
  cdm,
  conceptIds,
  gapEra,
  incidencePrevalenceCohortName,
  cohortDefinitionId = 1,
  overWrite = TRUE,
  verbose
)
}
\arguments{
\item{cdm}{cdm}

\item{conceptIds}{conceptIds}

\item{gapEra}{gapEra}

\item{incidencePrevalenceCohortName}{incidencePrevalenceCohortName}

\item{verbose}{verbose}
}
\description{
Get the cohorts to compute incidence and prevalence
}
