% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/largeScaleCharacterization.R
\name{largeScaleCharacterization}
\alias{largeScaleCharacterization}
\title{Explain function}
\usage{
largeScaleCharacterization(
  cdm,
  targetCohortName,
  targetCohortId = NULL,
  temporalWindows,
  tablesToCharacterize,
  characterizationTableName = "characterization",
  permanentCharacterizationTable = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{cdm}{cdm connection created with CDMConnector package}

\item{targetCohortName}{Name of the table in the cdm that contains the
target cohort}

\item{targetCohortId}{Cohort definition id for the analyzed target cohorts}

\item{temporalWindows}{Temporal windows that we want to characterize}

\item{tablesToCharacterize}{Name of the tables in the cdm that we want to
summarize}

\item{characterizationTableName}{Name of the table that contains the large
scale characterization}

\item{permanentCharacterizationTable}{if we want the table to be instantiated
as a permanent table or not}

\item{verbose}{Explain the steps of the program}
}
\description{
Explain function
}
