% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockDrugUtilisation.R
\name{mockDrugUtilisation}
\alias{mockDrugUtilisation}
\title{It creates a mock database for testing DrugUtilisation package}
\usage{
mockDrugUtilisation(
  numberIndividuals = 10,
  ...,
  source = "local",
  con = lifecycle::deprecated(),
  writeSchema = lifecycle::deprecated(),
  seed = lifecycle::deprecated()
)
}
\arguments{
\item{numberIndividuals}{Number of individuals in the mock cdm.}

\item{...}{Tables to use as basis to create the mock. If some tables are
provided they will be used to construct the cdm object.}

\item{source}{Source for the mock cdm, it can either be 'local' or 'duckdb'.}

\item{con}{deprecated.}

\item{writeSchema}{deprecated.}

\item{seed}{deprecated.}
}
\value{
A cdm reference with the mock tables
}
\description{
It creates a mock database for testing DrugUtilisation package
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

cdm
}

}
