% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{patternsWithFormula}
\alias{patternsWithFormula}
\title{Patterns and routes valid to compute daily dose}
\format{
A data frame with eight variables: \code{amount}, \code{amount_unit},
\code{numerator}, \code{numerator_unit}, \code{denominator},
\code{denominator_unit}, \code{route} and \code{formula_id}. See the
formulas using seeFormulas.
}
\usage{
patternsWithFormula
}
\description{
Patterns and routes valid to compute daily dose
}
\keyword{datasets}
