% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotTreatment}
\alias{plotTreatment}
\title{Generate a custom ggplot2 from a summarised_result object generated with
summariseTreatment function.}
\usage{
plotTreatment(
  result,
  facetX = "window_name",
  facetY = c("cdm_name", "cohort_name", "strata"),
  splitStrata = TRUE,
  colour = "treatment"
)
}
\arguments{
\item{result}{...}

\item{facetX}{...}

\item{facetY}{...}

\item{splitStrata}{...}

\item{colour}{...}
}
\value{
A ggplot2 object.
}
\description{
Generate a custom ggplot2 from a summarised_result object generated with
summariseTreatment function.
}
\examples{
\donttest{
library(duckdb)
library(CDMConnector)
library(CodelistGenerator)
library(PatientProfiles)
library(DrugUtilisation)

con <- dbConnect(duckdb(), eunomiaDir("GiBleed"))
cdm <- cdmFromCon(con = con, cdmSchema = "main", writeSchema = "main")

codes <- getDrugIngredientCodes(cdm = cdm, name = "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(
  cdm = cdm, conceptSet = codes, name = "study_cohort")

alternative <- getDrugIngredientCodes(cdm = cdm, name = c(
  "acetaminophen", "dextromethorphan", "doxylamine", "hydrocodone"))

result <- cdm$study_cohort |>
  addSex() |>
  addAge(ageGroup = list("<40" = c(0, 39), ">=40" = c(40, 150))) |>
  summariseTreatmentFromConceptSet(
    treatmentConceptSet = alternative,
    window = list(c(0, 199), c(200, 399), c(400, 599), c(600, 799)),
    strata = list(c("age_group", "sex"))
  )

plotTreatment(result)
}

}
