% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateConceptCohortSet.R,
%   R/summariseCharacteristics.R, R/summariseLargeScaleCharacteristics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{generateConceptCohortSet}
\alias{summariseCharacteristics}
\alias{summariseLargeScaleCharacteristics}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{CDMConnector}{\code{\link[CDMConnector]{generateConceptCohortSet}}}

  \item{PatientProfiles}{\code{\link[PatientProfiles]{summariseCharacteristics}}, \code{\link[PatientProfiles]{summariseLargeScaleCharacteristics}}}
}}

