% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{cohortCount}
\alias{settings}
\alias{attrition}
\alias{cohortCodelist}
\alias{bind}
\alias{suppress}
\alias{exportSummarisedResult}
\alias{mockDisconnect}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{omopgenerics}{\code{\link[omopgenerics]{attrition}}, \code{\link[omopgenerics]{bind}}, \code{\link[omopgenerics]{cohortCodelist}}, \code{\link[omopgenerics]{cohortCount}}, \code{\link[omopgenerics]{exportSummarisedResult}}, \code{\link[omopgenerics]{settings}}, \code{\link[omopgenerics]{suppress}}}

  \item{PatientProfiles}{\code{\link[PatientProfiles]{mockDisconnect}}}
}}

