% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require.R
\name{requirePriorDrugWashout}
\alias{requirePriorDrugWashout}
\title{Require prior days with no exposure.}
\usage{
requirePriorDrugWashout(
  cohort,
  priorUseWashout,
  cohortId = NULL,
  name = omopgenerics::tableName(cohort)
)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{priorUseWashout}{The length of priorUseWashout to be applied. NOTE that
if priorUseWashout is infinity requireIsFirstDrugEntry will be called
instead.}

\item{cohortId}{IDs of the cohorts to modify. The default is NULL meaning all
cohorts will be used; otherwise, only the specified cohorts will be modified,
and the rest will remain unchanged.}

\item{name}{Name of the new cohort. Default name is the original cohort name.}
}
\value{
The cohort table object with the requirements applied.
}
\description{
Require prior days with no exposure.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()
cdm$cohort1 <- cdm$cohort1 |>
  requirePriorDrugWashout(priorUseWashout = 90, cohortId = c(2,3))
attrition(cdm$cohort1)
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
