% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirePriorUseWashout.R
\name{requirePriorUseWashout}
\alias{requirePriorUseWashout}
\title{Generates a cohort of the drug use of a certain list of concepts.}
\usage{
requirePriorUseWashout(cohort, washout, name = tableName(cohort))
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{washout}{The length of washout to be applied.}

\item{name}{Name of the new cohort with the washout. Default name is the original
cohort name.}
}
\value{
The function returns the cohort after applying the specified washout.
}
\description{
Generates a cohort of the drug use of a certain list of concepts.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()
requirePriorUseWashout(cohort = cdm$cohort1, washout = 90)
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
