% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requirePriorUseWashout.R
\name{requirePriorUseWashout}
\alias{requirePriorUseWashout}
\title{Generates a cohort of the drug use of a certain list of concepts.}
\usage{
requirePriorUseWashout(cohort, washout)
}
\arguments{
\item{cohort}{A cohort table in a cdm reference.}

\item{washout}{The length of washout to be applied.}
}
\value{
The function returns the cohort after applying the specified washout.
}
\description{
Generates a cohort of the drug use of a certain list of concepts.
}
\examples{
\donttest{
cdm <- mockDrugUtilisation()
requirePriorUseWashout(cohort = cdm$cohort1, washout = 90)
CDMConnector::cdmDisconnect(cdm = cdm)
}

}
