% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern.R
\name{stratifyByUnit}
\alias{stratifyByUnit}
\title{Function to stratify a conceptSet by unit}
\usage{
stratifyByUnit(conceptSet, cdm, ingredientConceptId)
}
\arguments{
\item{conceptSet}{List of concept sets}

\item{cdm}{cdm reference}

\item{ingredientConceptId}{ConceptId that refers to an ingredient}
}
\value{
The conceptSet stratified by unit
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

codelist <- CodelistGenerator::getDrugIngredientCodes(cdm, "acetaminophen")

codelistStratified <- stratifyByUnit(codelist, cdm, 1125315)

codelistStratified
}

}
