% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseCharacteristics.R
\name{summariseCharacteristics}
\alias{summariseCharacteristics}
\title{This function has been deprecated please use CohortCharacteristics package}
\usage{
summariseCharacteristics()
}
\value{
An error pointing to the new function in CohortCharacteristics
}
\description{
This function has been deprecated please use CohortCharacteristics package
}
