% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseDoseTable.R
\name{summariseDoseTable}
\alias{summariseDoseTable}
\title{This function is used to summarise the dose table over multiple cohorts.}
\usage{
summariseDoseTable(
  cdm,
  strataCohortName,
  cohortId = NULL,
  doseTableName,
  variables = NULL,
  estimates = c("min", "max", "mean", "std", "median", "iqr", "q25", "q75"),
  minimumCellCounts = 5
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain the 'doseTableName' table and  also 'strataCohortName' and
'indicationList' if specified. It is a compulsory input, no default
value is provided.}

\item{strataCohortName}{Name to point to the cdm table that contains the
strata information. To compute this cohort you can use the getStratification
function or use any other function that the output is a cohort. It is a
compulsory input, no default value is provided.}

\item{cohortId}{cohort definition id of the cohorts in the strata cohort
table that you want to summarise. By default cohortId = NULL, then all
cohorts in strataCohort will be summarised.}

\item{doseTableName}{Name to point to the cdm table that contains the dose
cohort information computed using getDoseInformationTable}

\item{variables}{Vector with the variables that we want to summarise. The
variables must be contained in doseTableName. If NULL all variables in
doseTableName are summarised. By defaukt: NULL.}

\item{estimates}{Vector with the names of the functions used to summarise the
variables. The possible estimates are: "min", "max", "mean", "median", "iqr",
"range", "q5", "q10", "q15", "q20", "q25", "q30", "q35", "q40", "q45", "q50",
"q55", "q60", "q65", "q70", "q75", "q80", "q85", "q90", "q95" and "std". By
default: c("min", "max", "mean", "std", "median", "iqr", "q25", "q75").}

\item{minimumCellCounts}{Minimum counts that a group can have. Cohorts with
less counts than this value are obscured. By default: 5.}
}
\value{
A Tibble with 4 columns: cohort_definition_id, variable, estimate and
value. There will be one row for each cohort, variable and cohort
combination.
}
\description{
This function is used to summarise the dose table over multiple cohorts.
}
