% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseDrugUse.R
\name{summariseDrugUse}
\alias{summariseDrugUse}
\title{This function is used to summarise the dose table over multiple cohorts.}
\usage{
summariseDrugUse(
  cohort,
  cdm = attr(cohort, "cdm_reference"),
  strata = list(),
  drugUseVariables = drugUseColumns(cohort),
  drugUseEstimates = c("median", "q25", "q75"),
  minCellCount = 5
)
}
\arguments{
\item{cohort}{Cohort with drug use variables and strata}

\item{cdm}{cdm_reference generated by CDMConnector}

\item{strata}{Stratification list}

\item{drugUseVariables}{Name of columns with drug use variables}

\item{drugUseEstimates}{Estimates that we want for the columns}

\item{minCellCount}{Below this number counts will be supressed}
}
\value{
A summary of the drug use stratified by cohort_name and strata_name
}
\description{
This function is used to summarise the dose table over multiple cohorts.
}
\examples{
\donttest{
library(DrugUtilisation)
library(PatientProfiles)
library(CodelistGenerator)

concept_relationship <- dplyr::tibble(
concept_id_1 = c(2905077, 1516983, 2905075, 1503327, 1516978, 1503326, 1503328, 1516980,
                 29050773, 1125360, 15033297, 15030327, 15033427, 15036327, 15394662,
                 43135274, 11253605, 431352774, 431359274, 112530, 1539465, 29050772,
                 431352074, 15394062, 43135277, 15033327, 11253603, 15516980, 5034327,
                 1539462, 15033528, 15394636, 15176980, 1539463, 431395274, 15186980,
                 15316978),
concept_id_2 = c(19016586, 46275062, 35894935, 19135843, 19082107, 19011932, 19082108,
                 2008660,  2008661,  2008662, 19082109, 43126087, 19130307, 42629089,
                 19103220, 19082048, 19082049, 19082256, 19082050, 19082071, 19082072,
                 19135438, 19135446, 19135439, 19135440, 46234466, 19082653, 19057400,
                 19082227, 19082286, 19009068, 19082628, 19082224, 19095972, 19095973,
                 35604394, 702776 ),
relationship_id = c(rep("RxNorm has dose form", 37))
)
cdm <- mockDrugUtilisation(extraTables = list("concept_relationship" = concept_relationship))
cdm <- generateDrugUtilisationCohortSet(
  cdm, "dus_cohort", getDrugIngredientCodes(cdm, "acetaminophen")
)
cdm$dus_cohort <- cdm$dus_cohort \%>\%
  addDrugUse(cdm, 1125315)
result <- summariseDrugUse(cdm$dus_cohort, cdm)
print(result)

cdm$dus_cohort <- cdm$dus_cohort \%>\%
  addSex(cdm) \%>\%
  addAge(cdm, ageGroup = list("<40" = c(0, 30), ">40" = c(40, 150)))
result <- summariseDrugUse(
  cdm$dus_cohort, cdm, strata = list(
   "age_group" = "age_group", "sex" = "sex",
   "age_group and sex" = c("age_group", "sex")
  )
)
print(result)
}

}
