% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseIndication.R
\name{summariseIndication}
\alias{summariseIndication}
\title{This function is used to summarise the indication table over multiple
cohorts.}
\usage{
summariseIndication(
  cohort,
  strata = list(),
  indicationCohortName,
  indicationCohortId = NULL,
  indicationWindow = list(c(0, 0)),
  unknownIndicationTable = NULL
)
}
\arguments{
\item{cohort}{Cohort with indications and strata.}

\item{strata}{Stratification list.}

\item{indicationCohortName}{Name of indication cohort table}

\item{indicationCohortId}{target cohort Id to add indication}

\item{indicationWindow}{time window of interests}

\item{unknownIndicationTable}{Tables to search unknown indications}

\item{indexDate}{Date of the indication}
}
\value{
A summarise result object
}
\description{
This function is used to summarise the indication table over multiple
cohorts.
}
\examples{
\donttest{
library(DrugUtilisation)
library(PatientProfiles)
library(CDMConnector)

cdm <- mockDrugUtilisation()
indications <- list("headache" = 378253, "asthma" = 317009)
cdm <- generateConceptCohortSet(cdm, indications, "indication_cohorts")
acetaminophen <- getDrugIngredientCodes(cdm, "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(cdm, "drug_cohort", acetaminophen)
cdm[["drug_cohort"]] <- cdm[["drug_cohort"]] |>
  addIndication(
    indicationCohortName = "indication_cohorts",
    indicationWindow = list(c(0,0),c(-30,0),c(-365,0))
  )

}

}
