% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseIndication.R
\name{summariseIndication}
\alias{summariseIndication}
\title{This function is used to summarise the indication table over multiple
cohorts.}
\usage{
summariseIndication(
  cohort,
  cdm = lifecycle::deprecated(),
  strata = list(),
  minCellCount = lifecycle::deprecated()
)
}
\arguments{
\item{cohort}{Cohort with indications and strata.}

\item{cdm}{Deprecated.}

\item{strata}{Stratification list.}

\item{minCellCount}{Deprecated.}
}
\value{
A Tibble with 4 columns: cohort_definition_id, variable, estimate and
value. There will be one row for each cohort, variable and cohort
combination.
}
\description{
This function is used to summarise the indication table over multiple
cohorts.
}
\examples{
\donttest{
library(DrugUtilisation)
library(PatientProfiles)
library(CDMConnector)

cdm <- mockDrugUtilisation()
indications <- list("headache" = 378253, "asthma" = 317009)
cdm <- generateConceptCohortSet(cdm, indications, "indication_cohorts")
acetaminophen <- getDrugIngredientCodes(cdm, "acetaminophen")
cdm <- generateDrugUtilisationCohortSet(cdm, "drug_cohort", acetaminophen)
cdm[["drug_cohort"]] <- cdm[["drug_cohort"]] |>
  addIndication(
    indicationCohortName = "indication_cohorts",
    indicationWindow = list(c(0,0),c(-30,0),c(-365,0))
  )

}

}
