% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseIndication.R
\name{summariseIndication}
\alias{summariseIndication}
\title{This function is used to summarise the indication table over multiple
cohorts.}
\usage{
summariseIndication(
  cohort,
  cdm,
  strata = list(),
  indicationVariables = indicationColumns(cohort),
  minimumCellCount = 5
)
}
\arguments{
\item{cohort}{Cohort with indications and strata}

\item{cdm}{cdm_reference created by CDMConnector}

\item{strata}{Stratification list}

\item{indicationVariables}{Variables that point to an indication column}

\item{minimumCellCount}{Minimum counts that a group can have. Cohorts with
less counts than this value are obscured. By default: 5.}
}
\value{
A Tibble with 4 columns: cohort_definition_id, variable, estimate and
value. There will be one row for each cohort, variable and cohort
combination.
}
\description{
This function is used to summarise the indication table over multiple
cohorts.
}
