% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseDoseTable.R
\name{summariseIndication}
\alias{summariseIndication}
\title{This function is used to summarise the indication table over multiple
cohorts.}
\usage{
summariseIndication(
  cdm,
  cohortId = NULL,
  indicationList = NULL,
  minimumCellCounts = 5
)
}
\arguments{
\item{cdm}{'cdm' object created with CDMConnector::cdm_from_con(). It must
must contain the 'doseTableName' table and  also 'strataCohortName' and
'indicationList' if specified. It is a compulsory input, no default
value is provided.}

\item{cohortId}{cohort definition id of the cohorts in the strata cohort
table that you want to summarise. By default cohortId = NULL, then all
cohorts in strataCohort will be summarised.}

\item{indicationList}{is a list of tables that contains the indication
information. You must use the getIndication function to obtain this list. No
default value is provided.}

\item{minimumCellCounts}{Minimum counts that a group can have. Cohorts with
less counts than this value are obscured. By default: 5.}
}
\value{
A Tibble with 4 columns: cohort_definition_id, variable, estimate and
value. There will be one row for each cohort, variable and cohort
combination.
}
\description{
This function is used to summarise the indication table over multiple
cohorts.
}
