% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseProportionOfPatientsCovered.R
\name{summariseProportionOfPatientsCovered}
\alias{summariseProportionOfPatientsCovered}
\title{Summarise proportion Of patients covered}
\usage{
summariseProportionOfPatientsCovered(
  cohort,
  cohortId = NULL,
  strata = list(),
  followUpDays = NULL
)
}
\arguments{
\item{cohort}{A cohort table}

\item{cohortId}{Cohort definition ID of interest. If NUll, results for all
cohorts will be returned.}

\item{strata}{List of variables to stratify by.}

\item{followUpDays}{Number of days to follow up individuals for. If NULL the
maximum amount of days from an individuals first cohort start date to their
last cohort end date will be used}
}
\value{
A summarised result
}
\description{
Gives the proportion of patients still in observation who are
in the cohort on any given day following their first cohort entry. This is
known as the “proportion of patients covered” (PPC) method for assessing
treatment persistence.
}
