% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseIntersect.R
\name{summariseTreatment}
\alias{summariseTreatment}
\title{This function is used to summarise treatments received}
\usage{
summariseTreatment(
  cohort,
  window,
  treatmentCohortName,
  cohortId = NULL,
  treatmentCohortId = NULL,
  strata = list(),
  indexDate = "cohort_start_date",
  censorDate = NULL,
  mutuallyExclusive = FALSE,
  inObservation = TRUE,
  restrictIncident = FALSE
)
}
\arguments{
\item{cohort}{A cohort_table object.}

\item{window}{Time windows over which to identify treatments.}

\item{treatmentCohortName}{Name of the cohort table containing the treatments
of interest.}

\item{cohortId}{A cohort definition id to restrict by. If NULL, all cohorts
will be included.}

\item{treatmentCohortId}{Cohort definition IDs of the treatments of interest.
If \code{NULL}, all cohorts in \code{treatmentCohortName} are included.}

\item{strata}{A list of variables to stratify results. These variables
must have been added as additional columns in the cohort table.}

\item{indexDate}{Name of a column that indicates the date to start the
analysis.}

\item{censorDate}{Name of a column that indicates the date to stop the
analysis, if NULL end of individuals observation is used.}

\item{mutuallyExclusive}{Whether intersections should be mutually exclusive.
If \code{TRUE}, cohort combinations are reported as mutually exclusive categories;
if \code{FALSE}, each cohort is reported independently.}

\item{inObservation}{Whether to restrict the analysis to individuals in
observation. If \code{TRUE}, individuals not in observation are excluded. If
\code{FALSE}, they are included as a separate category.}

\item{restrictIncident}{Whether to include only target cohort records that
start during the analysis period. If \code{FALSE}, all target cohort records
that overlap with the study period will be included.}
}
\value{
A summary of treatments stratified by cohort_name and strata_name
}
\description{
This function is used to summarise treatments received
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()
cdm$cohort1 |>
  summariseTreatment(
    treatmentCohortName = "cohort2",
    window = list(c(0, 30), c(31, 365))
  )
}

}
