% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summariseTreatment.R
\name{summariseTreatmentFromConceptSet}
\alias{summariseTreatmentFromConceptSet}
\title{This function is used to summarise the dose table over multiple cohorts.}
\usage{
summariseTreatmentFromConceptSet(
  cohort,
  window,
  treatmentConceptSet,
  strata = list(),
  indexDate = "cohort_start_date",
  censorDate = NULL,
  combination = FALSE,
  minCellCount = lifecycle::deprecated()
)
}
\arguments{
\item{cohort}{Cohort with drug use variables and strata.}

\item{window}{Window where to summarise the treatments.}

\item{treatmentConceptSet}{Concept set list to summarise.}

\item{strata}{Stratification list.}

\item{indexDate}{Variable in x that contains the date to compute the
intersection.}

\item{censorDate}{Whether to censor overlap events at a specific date or a
column date of x. If NULL, end of observation will be used.}

\item{combination}{Whether to include combination treatments.}

\item{minCellCount}{\verb{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}}
}
\value{
A summary of the drug use stratified by cohort_name and strata_name
}
\description{
This function is used to summarise the dose table over multiple cohorts.
}
