% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatmentDiscontinuation.R
\name{treatmentDiscontinuation}
\alias{treatmentDiscontinuation}
\title{Compute the treatment dicontinuation.}
\usage{
treatmentDiscontinuation(
  cohort,
  cohortId = NULL,
  strata = NULL,
  censorOnDate = NULL,
  followUpDays = Inf,
  timeGap = 100,
  minCellCount = 5
)
}
\arguments{
\item{cohort}{GeneratedCohortSet object.}

\item{cohortId}{cohort_definition_id to target.}

\item{strata}{List of columns or combinations of columns to stratify for.}

\item{censorOnDate}{f not NULL, an individual's follow up will be censored at
the given date}

\item{followUpDays}{Number of days to follow up individuals (lower bound 1,
upper bound Inf)}

\item{timeGap}{Days between time points for which to report survival
estimates. First day will be day zero with risk estimates provided for times
up to the end of follow-up, with a gap in days equivalent to timeGap.}

\item{minCellCount}{The minimum number of events to reported, below which
results will be obscured. If 0, all results will be reported.}
}
\description{
Compute the treatment dicontinuation.
}
\examples{
\donttest{
library(DrugUtilisation)

cdm <- mockDrugUtilisation()

cdm <- generateDrugUtilisationCohortSet(
  cdm = cdm, name = "cohort", conceptSet = list("acetaminophen" = 1214)
)

treatmentDiscontinuation(cdm$cohort)
}

}
